remove unnecessary iteration over almost all items in the dep graph instead just check and emit this attribute error through `check_attr.rs`
