+++ compiler/rustc_symbol_mangling/src/lib.rs
@@ -106,7 +106,6 @@ mod legacy;
 mod v0;
 
 pub mod errors;
-pub mod test;
 
 pub use v0::mangle_internal_symbol;
